refactoring serialize.py
