+++ osgar/drivers/test_bus.py
@@ -7,6 +7,8 @@ from osgar.drivers.bus import (BusHandler, BusShutdownException,
                                LogBusHandler, LogBusHandlerInputsOnly,
                                serialize)
 
+from osgar.lib.serialize import serialize, deserialize
+
 
 class BusHandlerTest(unittest.TestCase):
 
@@ -115,11 +117,5 @@ class BusHandlerTest(unittest.TestCase):
         self.assertEqual(bus.listen(), (timedelta(microseconds=10), 'raw', [1, 2]))
         bus.publish('new_channel', b'some data')
         self.assertEqual(bus.listen(), (timedelta(microseconds=11), 'raw', [3, 4, 5]))
-
-    def test_serialization(self):
-            self.assertEqual(serialize(b'\x01\x02'), b'\xc4\x02\x01\x02')
-            position = (51749517, 180462688)
-            self.assertEqual(serialize(position), b'\x92\xce\x03\x15\xa2\x8d\xce\n\xc1\xa4`')
-            self.assertEqual(serialize((123.4, 'Hi')), b'\x92\xcb@^\xd9\x99\x99\x99\x99\x9a\xa2Hi')
 
 # vim: expandtab sw=4 ts=4
