+++ osgar/lib/serialize.py
@@ -1,0 +1,13 @@
+"""
+  serialization of messages
+"""
+import msgpack
+
+
+def serialize(data):
+    return msgpack.packb(data, use_bin_type=True)
+
+
+def deserialize(bytes_data):
+    return msgpack.unpackb(bytes_data, raw=False)
+
